Clean up ljson.hpp Static functions in a header are very strange so cleaned this up a bit to make this header actually a somewhat-desirable include
